pyproject.toml
[project]
name = "get-location"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "fake-useragent>=2.1.0",
    "geopy>=2.4.1",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "selenium>=4.30.0",
    "sqlalchemy>=2.0.40",
    "tqdm>=4.67.1",
]
